refactor: Update UnitUtil to return CGFloat COOEE-116
1. Usually UIView width, height are in CGFloat.
2. Rather than converting to GCFloat every time made all method return CGFloat
3. Before returning every converted pixel, converted it to points

+++ CooeeSDK/models/trigger/blocks/Border.swift
@@ -29,19 +29,19 @@ struct Border: HandyJSON {
     private var colour: Colour?
     private var style: Style?
 
-    public func getWidth(_ parent: UIView) -> Int {
+    public func getWidth(_ parent: UIView) -> CGFloat {
         return width != nil ? UnitUtil.getCalculatedValue(parent, width!) : 0
     }
 
-    public func getRadius(_ parent: UIView) -> Int {
+    public func getRadius(_ parent: UIView) -> CGFloat {
         return radius != nil ? UnitUtil.getCalculatedPixel(radius!) : 0
     }
 
-    public func getDashWidth(_ parent: UIView) -> Int {
+    public func getDashWidth(_ parent: UIView) -> CGFloat {
         return dashWidth != nil ? UnitUtil.getCalculatedPixel(dashWidth!) : 0
     }
 
-    public func getDashGap(_ parent: UIView) -> Int {
+    public func getDashGap(_ parent: UIView) -> CGFloat {
         return dashGap != nil ? UnitUtil.getCalculatedPixel(dashGap!) : 0
     }
